feat(voyager): handle truly fatal errors
certain errors (such as a counterparty chain not existing in the config) are
truly fatal in that they will never be recoverable - they will just cause a
restart loop. in this case, we need to actually handle the error and propagate
it up to `queue_msg`, where it can be marked as failed and execution can continue.